feat(ProficiencyChart): extract interfaces to a separate file for better organization
